projects/dailytunztunz/server/requirements.txt
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pycodestyle==2.5.0
python-dotenv==0.10.3
requests==2.22.0
urllib3==1.25.6
Werkzeug==0.16.0
